ooDialog - fix warnings on 64-bit
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@6069 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/oodPropertySheetDialog.cpp
@@ -307,7 +307,7 @@ INT_PTR getSetActiveValue(RexxThreadContext *c, pCPropertySheetDialog pcpsd, Rex
         return (INT_PTR)c->PointerValue((RexxPointerObject)result);
     }
 
-    int max = pcpsd->pageCount;
+    int max = (int)pcpsd->pageCount;
 
     int32_t index;
     if ( ! c->Int32(result, &index) || (index < -1 || index > max) )
@@ -1980,7 +1980,7 @@ RexxMethod2(RexxObjectPtr, psdlg_indexToID, int32_t, index, CSELF, pCSelf)
 
     INT_PTR result = -2;
 
-    int max = pcpsd->pageCount;
+    int max = (int)pcpsd->pageCount;
 
     if ( index < -1 || index > max )
     {
